+++ tasmota/language/en_GB.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Other parameters"
 #define D_TEMPLATE "Template"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Activate"
 #define D_DEVICE_NAME "Device Name"
 #define D_WEB_ADMIN_PASSWORD "Web Admin Password"