+++ tasmota/language/es_ES.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Otros parámetros"
 #define D_TEMPLATE "Plantilla"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Activar"
 #define D_DEVICE_NAME "Nombre de Dispositivo"
 #define D_WEB_ADMIN_PASSWORD "Clave Administrador Web"