+++ tasmota/language/fr_FR.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Autres paramètres"
 #define D_TEMPLATE "Modèle"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Activer"
 #define D_DEVICE_NAME "Nom de l'appareil"
 #define D_WEB_ADMIN_PASSWORD "Mot de passe Web Admin"