+++ tasmota/language/fy_NL.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Oare parameters"
 #define D_TEMPLATE "Sjabloan"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Aktivearje"
 #define D_DEVICE_NAME "Namme fan it apparaat"
 #define D_WEB_ADMIN_PASSWORD "Web Admin Wachtwurd"