+++ tasmota/language/hu_HU.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Egyéb beállítások"
 #define D_TEMPLATE "Template"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Aktiválás"
 #define D_DEVICE_NAME "Eszköz neve"
 #define D_WEB_ADMIN_PASSWORD "Web admin jelszó"