+++ tasmota/language/it_IT.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS    "Altri parametri"
 #define D_TEMPLATE            "Modello"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE            "Attiva"
 #define D_DEVICE_NAME         "Nome dispositivo"
 #define D_WEB_ADMIN_PASSWORD  "Password amministratore web"