+++ tasmota/language/nl_NL.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Overige parameters"
 #define D_TEMPLATE "Sjabloon"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Activeer"
 #define D_DEVICE_NAME "Apparaatnaam"
 #define D_WEB_ADMIN_PASSWORD "Web Admin Wachtwoord"