+++ tasmota/language/pt_BR.h
@@ -349,6 +349,7 @@
 
 #define D_OTHER_PARAMETERS "Outros parâmetros"
 #define D_TEMPLATE "Modelo"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Ativar"
 #define D_DEVICE_NAME "Nome do dispositivo"
 #define D_WEB_ADMIN_PASSWORD "Senha de WEB Admin"