+++ tasmota/language/ru_RU.h
@@ -350,6 +350,7 @@
 
 #define D_OTHER_PARAMETERS "Настройки Прочие"
 #define D_TEMPLATE "Шаблон"
+#define D_TEMPLATE_WRONG_ARCH "Reset to correct chip type"
 #define D_ACTIVATE "Активировать"
 #define D_DEVICE_NAME "Имя устройства"
 #define D_WEB_ADMIN_PASSWORD "Пароль веб-интерфейса"