+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Support for Sensirion SCD43 CO2 sensor
 - I2S full duplex, auto rx sample rate (#24469)
 - Support for Sensirion STCC4 CO2 sensor
+- Support for JSON value pair `"ARCH"` in template being either ESP8266, ESP32, ESP32C2, ESP32C3, ESP32C5, ESP32C6, ESP32H2, ESP32H4, ESP32P4, ESP32S2 or ESP32S3
 
 ### Breaking Changed
 
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - Do not free BT memory when in use (#24480)
 - Berry avoid `tasmota.wifi()` returning bad values when wifi is turned off (#24505)
 - Don't send extraneous `0\r\n\r\n` with non-chunked HTTP/1.0 (#24518)
+- File upload improvements: `/ufsu` api mode, no interrupts disabling, cleaner confirmation page (#24521)
 
 ### Removed
 - Berry `tasmota.urlbecload()` superseded by Extension Manager (#24493)